Just share the u64 directly from Zig

+++ AdventOfCode.Solutions.Zig/src/timer.zig
@@ -8,16 +8,16 @@ const Key = struct {
     part: core.Part,
 };
 
-var durations: ?*std.AutoHashMap(Key, i128) = null;
+var durations: ?*std.AutoHashMap(Key, u64) = null;
 
 var init_timer_once = std.once(initializeTimer);
 
 fn initializeTimer() void {
-    durations = std.heap.page_allocator.create(std.AutoHashMap(Key, i128)) catch return;
-    durations.?.* = std.AutoHashMap(Key, i128).init(std.heap.page_allocator);
+    durations = std.heap.page_allocator.create(std.AutoHashMap(Key, u64)) catch return;
+    durations.?.* = std.AutoHashMap(Key, u64).init(std.heap.page_allocator);
 }
 
-pub fn getDuration(year: u16, day: u8, part: core.Part) ?i128 {
+pub fn getDuration(year: u16, day: u8, part: core.Part) ?u64 {
     init_timer_once.call();
 
     if (durations) |map| {
@@ -25,7 +25,7 @@ pub fn getDuration(year: u16, day: u8, part: core.Part) ?i128 {
     } else return null;
 }
 
-pub fn setDuration(year: u16, day: u8, part: core.Part, duration: i128) void {
+pub fn setDuration(year: u16, day: u8, part: core.Part, duration: u64) void {
     init_timer_once.call();
 
     if (durations) |durations_map| {
